feat: created empty prompt component and made it easier to read
