+++ src/components/EmptyPrompts/empty-prompts.module.scss
@@ -1,0 +1,23 @@
+@use 'sass:color';
+@import '@/variables.module';
+
+.emptyPersonalPrompts {
+  display: flex;
+  flex-direction: column;
+  gap: 16px;
+  width: 100%;
+  height: 164px;
+  padding: 30px;
+  justify-content: center;
+  align-items: center;
+  flex-shrink: 0;
+  border-radius: 20px;
+  border: 1px solid var(--Border-Tertiary, #222);
+
+  p {
+    color: $light-40;
+    font-size: 13px;
+    font-weight: 500;
+    line-height: 16px;
+  }
+}
