+++ src/components/PersonalPrompts/PersonalPrompts.tsx
@@ -1,172 +1,43 @@
 import styles from './personal-prompts.module.scss'
 import variables from '@/variables.module.scss'
 import { useState } from 'react'
+import { PersonalPromptsProps, PersonalPromptsItemProps } from '@/types'
+
 // Components
 import { Badge } from '@/components/Badge/Badge'
 import Button from '@/components/Button/Button'
 import { Setting, Trash, Lock, Edit2 } from 'iconsax-react'
-import { Prompt } from '@/types/supabase-helpers'
+import EmptyPrompts from '@/components/EmptyPrompts/EmptyPrompts'
 
-const CSS_VARIABLES = {
-  private: {
-    icon: {
-      color: variables.textPrimary,
-    },
-    background: {
-      color: variables.backgroundSecondary,
-      hoverColor: '#292929',
-    },
-    button: {
-      textColor: variables.textSecondary,
-      backgroundColor: variables.backgroundTertiary,
-      hoverBackgroundColor: variables.backgroundTertiary,
-      borderColor: variables.tertiary50,
-      hoverBorderColor: variables.tertiary50,
-    },
-    ctaButton: {
-      textColor: variables.textSecondary,
-      backgroundColor: variables.backgroundTertiary,
-      hoverBackgroundColor: variables.backgroundTertiary,
-      borderColor: variables.backgroundTertiary,
-      hoverBorderColor: variables.tertiary50,
-    },
-    useButton: {
-      textColor: variables.primary100,
-      backgroundColor: variables.backgroundTertiary,
-      hoverBackgroundColor: variables.backgroundTertiary,
-      borderColor: variables.backgroundTertiary,
-      hoverBorderColor: '#00EAFF70',
-    },
-  },
-  public: {
-    icon: {
-      color: variables.pink100,
-    },
-    background: {
-      color: variables.backgroundSecondary,
-      hoverColor: variables.pink20,
-    },
-    button: {
-      textColor: variables.textSecondary,
-      backgroundColor: variables.backgroundTertiary,
-      hoverBackgroundColor: variables.pink20,
-      borderColor: variables.backgroundTertiary,
-      hoverBorderColor: variables.pink40,
-    },
-    ctaButton: {
-      textColor: variables.textSecondary,
-      backgroundColor: variables.backgroundTertiary,
-      hoverBackgroundColor: variables.pink20,
-      borderColor: variables.backgroundTertiary,
-      hoverBorderColor: variables.pink40,
-    },
-    useButton: {
-      textColor: variables.textSecondary,
-      backgroundColor: variables.backgroundTertiary,
-      hoverBackgroundColor: variables.pink20,
-      borderColor: variables.backgroundTertiary,
-      hoverBorderColor: variables.pink40,
-    },
-  },
-  forked: {
-    icon: {
-      color: variables.primary100,
-    },
-    background: {
-      color: variables.backgroundSecondary,
-      hoverColor: variables.primary20,
-    },
-    button: {
-      textColor: variables.textSecondary,
-      backgroundColor: variables.backgroundTertiary,
-      hoverBackgroundColor: variables.primary20,
-      borderColor: variables.backgroundTertiary,
-      hoverBorderColor: variables.primary20,
-    },
-    ctaButton: {
-      textColor: variables.textSecondary,
-      backgroundColor: variables.backgroundTertiary,
-      hoverBackgroundColor: variables.primary20,
-      borderColor: variables.backgroundTertiary,
-      hoverBorderColor: variables.primary20,
-    },
-    useButton: {
-      textColor: variables.textSecondary,
-      backgroundColor: variables.backgroundTertiary,
-      hoverBackgroundColor: variables.primary20,
-      borderColor: variables.backgroundTertiary,
-      hoverBorderColor: variables.primary20,
-    },
-  },
-}
+// Custom Variables for styling
+import CSS_VARIABLES from './customVariables'
 
-const PersonalPrompts = ({
-  userId,
-  prompts,
-  openModal,
-  selectPrompt,
-}: {
-  userId: string | undefined
-  prompts: Prompt[]
-  openModal: (modal: string, context?: Record<string, any>) => void
-  selectPrompt: (prompt: Prompt) => void
-}) => {
+const PersonalPrompts = ({ userId, prompts, openModal, selectPrompt }: PersonalPromptsProps) => {
+  if (prompts.length === 0) {
+    return <EmptyPrompts openModal={openModal} />
+  }
+
   return (
     <div className={styles.personalPromptsContainer}>
-      {prompts.length > 0 ? (
-        <div className={styles.personalPromptsHeader}>
-          <h2>My Prompts</h2>
-          <div className={styles.personalPrompts}>
-            {prompts.map((item) => (
-              <PersonalPromptsItem
-                key={item.name}
-                userId={userId}
-                name={item.name}
-                description={item.description ?? ''}
-                slug={item.slug}
-                selectPrompt={selectPrompt}
-                prompt={item}
-                openModal={openModal}
-                externalId={item.external_id}
-                publicUseNumber={item.public_use_number}
-              />
-            ))}
-          </div>
+      <div className={styles.personalPromptsHeader}>
+        <h2>My Prompts</h2>
+        <div className={styles.personalPrompts}>
+          {prompts.map((item) => (
+            <PersonalPromptsItem
+              key={item.name}
+              userId={userId}
+              prompt={item}
+              selectPrompt={selectPrompt}
+              openModal={openModal}
+            />
+          ))}
         </div>
-      ) : (
-        <div className={styles.emptyPersonalPrompts}>
-          <Setting color={variables.light20} variant={'Bold'} />
-          <p>Prompts you create and favorite will be added here.</p>
-          <Button size="small" variant="tertiary" onClick={() => openModal('create-prompt')}>
-            Create Prompt
-          </Button>
-        </div>
-      )}
+      </div>
     </div>
   )
 }
 
-const PersonalPromptsItem = ({
-  userId,
-  name,
-  description,
-  slug,
-  selectPrompt,
-  prompt,
-  openModal,
-  externalId,
-  publicUseNumber,
-}: {
-  userId: string | undefined
-  name: string
-  description?: string
-  slug: string
-  selectPrompt: (prompt: Prompt) => void
-  prompt: Prompt
-  openModal: (modal: string, context?: Record<string, any>) => void
-  externalId: string
-  publicUseNumber: number
-}) => {
+const PersonalPromptsItem = ({ userId, prompt, selectPrompt, openModal }: PersonalPromptsItemProps) => {
   const [isCopied, setIsCopied] = useState(false)
   const [isHovered, setIsHovered] = useState(false)
 
@@ -191,11 +62,11 @@ const PersonalPromptsItem = ({
     >
       <div className={styles.personalPromptsItemHeader}>
         <Setting color={colorScheme.icon.color} variant={'Bulk'} />
-        <h3>{name}</h3>
+        <h3>{prompt.name}</h3>
       </div>
       <div className={styles.personalPromptsItemContent}>
-        {description ? (
-          <p>{description?.length > 90 ? `${description.substring(0, 90)}...` : description}</p>
+        {prompt.description ? (
+          <p>{prompt.description?.length > 90 ? `${prompt.description.substring(0, 90)}...` : prompt.description}</p>
         ) : (
           <p>No description</p>
         )}
@@ -215,7 +86,7 @@ const PersonalPromptsItem = ({
                 color: colorScheme.ctaButton.textColor,
               }}
               onClick={() => {
-                const url = `https://chat.highlight.ing/prompts/${slug}`
+                const url = `https://chat.highlight.ing/prompts/${prompt.slug}`
                 navigator.clipboard.writeText(url)
                 setIsCopied(true)
                 setTimeout(() => setIsCopied(false), 2000)
@@ -242,7 +113,7 @@ const PersonalPromptsItem = ({
             </Button>
           )}
           <Badge variant="disabled" hidden={isHovered}>
-            {publicUseNumber ? `${publicUseNumber} Uses` : 'No uses'}
+            {prompt.public_use_number ? `${prompt.public_use_number} Uses` : 'No uses'}
           </Badge>
           <Button
             className={styles.filledButton}
@@ -265,7 +136,7 @@ const PersonalPromptsItem = ({
             style={{
               border: `1px solid ${isHovered ? colorScheme.button.hoverBorderColor : colorScheme.button.borderColor}`,
             }}
-            onClick={() => openModal('confirm-delete-prompt', { externalId: externalId, name: name })}
+            onClick={() => openModal('confirm-delete-prompt', { externalId: prompt.external_id, name: prompt.name })}
             hidden={!isHovered}
           >
             <Trash color={variables.tertiary} variant={'Bold'} size="24" />
