+++ src/types/index.ts
@@ -1,3 +1,5 @@
+import { Prompt } from '@/types/supabase-helpers'
+
 type WindowAttachment = {
   type: 'window'
   title: string
@@ -157,4 +159,18 @@ export type DownloadPlatformType = 'windows' | 'mac-intel' | 'mac-silicon' | 'un
 export type PromptTag = {
   value: string
   label: string
+}
+
+export interface PersonalPromptsProps {
+  userId: string | undefined
+  prompts: Prompt[]
+  openModal: (modal: string, context?: Record<string, any>) => void
+  selectPrompt: (prompt: Prompt) => void
+}
+
+export interface PersonalPromptsItemProps {
+  userId: string | undefined
+  prompt: Prompt
+  selectPrompt: (prompt: Prompt) => void
+  openModal: (modal: string, context?: Record<string, any>) => void
 }
